+++ Pedidos/Controllers/FormaPagamentoController.cs
@@ -22,12 +22,20 @@ namespace Pedidos.Controllers
         // GET: FormaPagamento
         public async Task<IActionResult> Index()
         {
+            if (!ValidarCuenta())
+            {
+                return RedirectToAction("Salir", "Login");
+            }
             return View(await _context.P_FormaPagamento.Where(x => x.idCuenta == Cuenta.id && x.activo).ToListAsync());
         }
 
         // GET: FormaPagamento/Details/5
         public async Task<IActionResult> Details(int? id)
         {
+            if (!ValidarCuenta())
+            {
+                return RedirectToAction("Salir", "Login");
+            }
             if (id == null)
             {
                 return NotFound();
@@ -46,6 +54,10 @@ namespace Pedidos.Controllers
         // GET: FormaPagamento/Create
         public IActionResult Create()
         {
+            if (!ValidarCuenta())
+            {
+                return RedirectToAction("Salir", "Login");
+            }
             return View();
         }
 
@@ -56,6 +68,10 @@ namespace Pedidos.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("id,nombre,idCuenta,activo")] P_FormaPagamento p_FormaPagamento)
         {
+            if (!ValidarCuenta())
+            {
+                return RedirectToAction("Salir", "Login");
+            }
             if (ModelState.IsValid)
             {
                 p_FormaPagamento.idCuenta = Cuenta.id;
@@ -71,6 +87,10 @@ namespace Pedidos.Controllers
         // GET: FormaPagamento/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+            if (!ValidarCuenta())
+            {
+                return RedirectToAction("Salir", "Login");
+            }
             if (id == null)
             {
                 return NotFound();
@@ -91,6 +111,10 @@ namespace Pedidos.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("id,nombre,idCuenta,activo")] P_FormaPagamento p_FormaPagamento)
         {
+            if (!ValidarCuenta())
+            {
+                return RedirectToAction("Salir", "Login");
+            }
             if (id != p_FormaPagamento.id)
             {
                 return NotFound();
@@ -122,6 +146,10 @@ namespace Pedidos.Controllers
         // GET: FormaPagamento/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+            if (!ValidarCuenta())
+            {
+                return RedirectToAction("Salir", "Login");
+            }
             if (id == null)
             {
                 return NotFound();
@@ -142,6 +170,10 @@ namespace Pedidos.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (!ValidarCuenta())
+            {
+                return RedirectToAction("Salir", "Login");
+            }
             var p_FormaPagamento = await _context.P_FormaPagamento.FindAsync(id);
             _context.P_FormaPagamento.Remove(p_FormaPagamento);
             await _context.SaveChangesAsync();
